TASK-282 - Fix TUI detail pane loading state regression
